+++ cmd/frontend/main.go
@@ -33,7 +33,6 @@ func main() {
 
 	if err := grpcclient.RunFromEnvironment(ctx, mux.Handler(
 		// copy/paster's beware: [frontend.WithTargetForwardingHandler] should not be set except for the root dalec frontend.
-		// frontend.WithBuiltinHandler(azlinux.Mariner2TargetKey, azlinux.NewMariner2Handler()),
 		azlinux.Handlers,
 		frontend.WithBuiltinHandler(windows.DefaultTargetKey, windows.Handle),
 		ubuntu.Handlers,
