--- frontend/linux/rpm/distro/common.go
@@ -1,19 +1,0 @@
-package distro
-
-import (
-	"fmt"
-	"path/filepath"
-)
-
-func importGPGScript(keyPaths []string) string {
-	// all keys that are included should be mounted under this path
-	keyRoot := "/etc/pki/rpm-gpg"
-
-	var importScript string = "#!/usr/bin/env sh\nset -eux\n"
-	for _, keyPath := range keyPaths {
-		keyName := filepath.Base(keyPath)
-		importScript += fmt.Sprintf("gpg --import %s\n", filepath.Join(keyRoot, keyName))
-	}
-
-	return importScript
-}
